Refactor: Move identifier-case bootstrap
